requirements.txt
fastapi==0.134.0
uvicorn==0.41.0
google-genai==1.65.0
yt-dlp==2026.3.3
python-dotenv==1.2.1
python-multipart==0.0.22
google-api-python-client